:recycle: refactor : eventList 코드 리팩토링 - addClickListenerFor* -> click* 로 파일명 및 함수 변경 - eventList 페이지에서 today 아이콘 클릭시 호버효과 적용 - month의 clickToday에 있는 today 아이콘 호버 로직 함수화
