+++ views/eventList.ejs
@@ -1,5 +1,5 @@
 <%- contentFor('scriptFiles') %>
-<script defer type="module" src="/javascripts/eventList/setLayoutForEventList.js"></script>
+<script defer type="module" src="/javascripts/eventList/rendEventListPage.js"></script>
 <%- contentFor('content') %>
 <section class="eventlist"></section>
 <%- contentFor('eventForm') %> <%- include('components/eventForm') %> <%- contentFor('eventDetails')
